+++ packages/flutter_rbac_service_data_interface/lib/flutter_rbac_service_data_interface.dart
@@ -2,9 +2,10 @@
 library flutter_rbac_service_data_interface;
 
 export 'src/local_rbac_datasource.dart';
-export 'src/models/account_data_model.dart';
+export 'src/models/account_group_model.dart';
+export 'src/models/account_model.dart';
 export 'src/models/assignment_model.dart';
+export 'src/models/permission_group_model.dart';
 export 'src/models/permission_model.dart';
-export 'src/models/role_data_model.dart';
-export 'src/models/securable_object_data_model.dart';
+export 'src/models/securable_object_model.dart';
 export 'src/rbac_data_interface.dart';